graph-datasources/graph-earnings-announcement-transcripts/requirements.txt
# Core dependencies for graph extraction pipelines
# SEC 10-K Filings and Earnings Announcement Transcripts

# Core libraries
llama-index-core>=0.14.0
python-dotenv>=1.2.0
openai>=2.9.0

# Google Cloud Storage
google-cloud-storage>=3.7.0

# Graph database
GQLAlchemy>=1.8.0

# DragonFly/Redis client
redis>=5.0.0

# spaCy and transformer support
# Note: Install with --only-binary=:all: to avoid compilation issues on Python 3.13
# Example: pip install spacy==3.8.11 --only-binary=:all:
spacy==3.8.11
spacy-curated-transformers==0.3.1

# Dependencies (installed automatically, but listed for reference)
# blis==1.3.3
# thinc==8.3.10
# curated-transformers==0.1.1
# transformers==4.49.0

